chore: refactor FadeInTransition
