+++ src/core/client/admin/routes/Moderate/ModerateCard/ModerateCardContainer.tsx
@@ -12,6 +12,7 @@ import NotAvailable from "coral-admin/components/NotAvailable";
 import { getModerationLink } from "coral-admin/helpers";
 import { ApproveCommentMutation } from "coral-admin/mutations";
 import { RejectCommentMutation } from "coral-admin/mutations";
+import FadeInTransition from "coral-framework/components/FadeInTransition";
 import {
   MutationProp,
   withFragmentContainer,
@@ -19,7 +20,6 @@ import {
 } from "coral-framework/lib/relay";
 import { GQLTAG } from "coral-framework/schema";
 
-import FadeInTransition from "./FadeInTransition";
 import FeatureCommentMutation from "./FeatureCommentMutation";
 import ModerateCard from "./ModerateCard";
 import ModeratedByContainer from "./ModeratedByContainer";
