+++ src/core/client/framework/components/index.ts
@@ -4,4 +4,5 @@ export { default as FacebookButton } from "./FacebookButton";
 export { default as GoogleButton } from "./GoogleButton";
 export { default as OIDCButton } from "./OIDCButton";
 export { default as Markdown } from "./Markdown";
+export { default as FadeInTransition } from "./FadeInTransition";
 export { default as DurationField, DURATION_UNIT } from "./DurationField";
