--- src/core/client/stream/tabs/Comments/Stream/FadeInTransition.css
@@ -1,8 +1,0 @@
-.appear {
-  opacity: 0;
-  pointer-events: none;
-}
-.appearActive {
-  opacity: 1;
-  transition: opacity 400ms;
-}
